Add HQ request timeouts
Read timeout of 15 minutes matches the expiration period of locks used around restore requests to hopefully prevent those requests from hanging indefinitely and subsequently locking a user out of formplayer (until the next restart). While it's good to have this in place as a fallback, 15 minute lock periods make for a bad user experience, and this does not fix the root cause of that issue.


Former-commit-id: e66321cc7dbab050f7336ffcd381b7a575ea09b0