don't autowire the request
Previously this worked because we were creating a mock FormplayerHttpRequest
bean in the TestContext which would match as the autowire target
when being called outside of a request scope. Inside the request scope
the request provided by the test framework would get used.

With the removal of the custom FormplayerHttpRequest object there is now
no bean that matches the autowire target outside of a request cycle.

Instead of 'faking' it rather just check if the request is null since this
is only used for logging and metrics.