improved 'back' navigation after form submission
Attempt to navigate to the previous screen prior to form submission if the user attempts to (e.g. using browser's back button).

This is an improvement of the current functionality which will start a new form session or else error if the case that was selected is closed or the form is not longer visible.