Only fully init entity screen when we are showing entity screen to the user
Earlier we were creating the entity list as part of entity response class.Therefore we could afford to set needsFullInit to false in cases when there is no selection over the entity list. Though entity list creation has now moved to Entity Screen init itself from the response class and response class simply takes the entity list from the screen class. Therfore we need to make sure that entity screen builds the entity list when the response is going to be an entity list response.