logs instead of throwing exception where match is not found
The scenario where a match is not found is a valid error. However,
the pre-existing behavior is that if a match is not found, it navigates
to the last valid step. Apps have been built expecting this behavior.
As such, apps built with the expection to navigate to the last valid step,
were now encountering this error. So this error handling will be now log
instead of throwing an exception.